Allow any version of node-sass to be used with this plugin
I'd like to allow callers of the plugin to specify whichever version of node-sass they want, so that it's easier to pin
to an exact version or version range. This should also reduce the maintenance overhead for grunt-sass, since the plugin
no longer has to be updated every time there is a new node-sass (beta) release.

Express node-sass as both a peerDependency and devDependency, so that it's still possible to run the test suite. Prefer
non-beta version for the test suite, since the beta dist-tag seems to prevent the peer dependency from being satisfied
on npm v2.14.12.